BONE: Switch J and Q keys
Q is as very often used key for closing applications and should
therefore be better reachable.

Also this change brings the Q into its QWERTY position, which is nice.

The thirt advantage is that Q will be operated with the left hand, which
is helpful when the right hands is resting on the mouse.