Modify NUMPAD layer
The new NUMPAD layer is made to let the hand rest in the home position
for natural entering of numbers.

It resembles the usage of the numbers in the num-block of traditional
keyboards. The middle finger rests on the 5, the thumb rests on the 0.
Enter can be hit the joint of the right pinky. However the placement of
the mathematical operators is different and all operators are aligned in
a single row.

Left of the number block are keys for the separators:
 . (as a thousands separator like in Germany or a fractions separator in the US)
 , (the same as above, but reversed)
 : (as a separator for clock time values)

The Enter key and the Equls sign are mapped on multiple positions to allow for different habits.

The layout looks as following:

     = -
   :789+
   .456/=
   ,123*↵

  0↵