Add an UMLAUT layer
A new layer has been added for entering german Umlaut characters. The
position of the Umlauts matches the position of the corresponding vowels
in the BONE layer.

The layer shift key has been added to the BONE layer and made
one-shottable. This allows a handling similar to a Compose key, but with
much less keystrokes (only the layer shift key and the vowel).

This change allows for removing the Umlaut keys from the base (BONE)
layer and adding other keys in their position.