Add Ctrl-W and Ctrl-S to FUNCTION layer
Ctrl-W (Close Window) and Ctrl-S (Save) are often used shortcuts in a
broad range of applications. These shortcuts are now added to the
function layer to be accesible with the left hand (which they are in a
QWERTY/QWERTZ layout, but not in BONE) to make them easily reachable
while the right hand rests on the mouse.