Correct QWERTY for host side QWERTZ layout
All the keys that differ between QWERTY and QWERTZ had to be adapted to
produce the correct (QWERTY) output when being used with a host side
QWERTZ layout.

It is still not perfect, since the shifte values still don’t match. Some
Kaleidoscope plugins have to be configured to properly support them.