Add BONE layers
Replace the existing primary layer with a new one with a BONE [1]
layout. Also add a symbol layer from the BONE (and therefore the
NEO/ADNW) layout.

The original QWERTY layer is retained as a layer above BONE. It is
currently unreachable since the firmware in this commit has no mapping
to change to it. This will be added in a later commit.

[1] https://www.neo-layout.org/Layouts/bone/