BONE: Remove Backtick and add Capslock left outer
Backtick is provided by the SYMBOL layer now.

CapsLock should be at an easy to reach position as I use it as Compose
key. The same position as on a traditional kayboard seems to be the best
one for that purpose.

The position of the PageUp and PageDown keys is still intuitive for me.