Z and Period keys double as shift to layer SYMBOL
The Z and Period keys now also shift to the SYBMOL layer when held.

This allows entering symbols from the right side by holding the Period
key with the left index finger and symbols from the left side by holding
the Z key with the right index finger. Unfortunately it is hard to enter
symbols from the left and the right side directly after each other.