Define colors for the different layers
Define a simple EGA palette and assign colors to different keys.

1. BONE:     No colors
2. QWERTY:   Q and W in brown (to indicate that the layer is active)
3. NUMPAD:   Numbers red, Operators blu, Separators green, Enter yellow
4. FUNCTION: Navigation blue, Modification red, Mouse green, Media pink
5. SYMBOL:   All characters blue
6. UMLAUT:   Umlaute blue, j and ß green

It is not finished yet. Especially the SYMBOL layer could get more
colored keys to indicate which keys are actually mapped in that layer.
Also the colors themselves should be revised. The palette should contain
a bit nicer colors. The meaning of the colors should also be defined.