UMLAUT: Add ”qu” as macro
Add a macro to type “qu” at once since “q” is usually followed by “u” in
written text.

However this approach has a flaw. When pressing a (one-shot) shift
beforehand, /both/ characters get uppercased. This is not the case when
using Dynamic Macros from Chrysalis. So probably we have to use Dynamic
Macros here as well.