SYM: Add DoubleQuote to left side of keyboard
As the Asterisk is on the left side of the keyoard and is often used in
combination with the double quote (for example when using “find” in the
shell) it was now added to the left side of the keyboard as well.

It was deliberately put near the Asterisk and therefore the less often
used Caret had to be moved to a still free position on the keymap.