Change expected host layout to QWERTY (US intl)
The symbolic key names are therefore changed back to the original ones.
Additional symbolic names have been added to hupfdule-keys.h and used in
the keymaps.

It is not perfect yet as some keys have different shifted chars. That is
mainly the comma and the period key that should produce a semicolon and
a colon in their shifted stated, but now produce a less-than and a
greater-than char (like in QWERTY).

The Caret (^) and Backtick (`) characters have been changed to expect
non-dead keys. That requires a change in the hosts keymap, since (at
least under Windows) those keys are dead keys in the US intl layout.
Under Linux it is possible to specify “nodeadkeys” as an option to
setxkbmap.