Change shifted versions of , and . to ; and :
Due to the change from a german host keymap to an english one some keys
need to be adapted to still produce the same characters when shift is
being pressed.

Due to technical limitations of the used plugins (CharShift and Qukeys)
this requires that we configure the dual use keys (Qukeys)
programmatically rather than declaratively in the keymaps. Therefore the
keymap now only contains the base key and Qukeys as well as CharShift is
configured in the setup() method.