Move dual-use layer shift to SYMBOL to different keys
The “.” and “z” keys were uses as dual use keys before, but that leads
to a high amount of mistyping due to the misinterpretation of primary
and secondary mappings on those keys. By moving that dual-use
functionality to the rarely used keys “Tab” and “ß" this problem can be
avoided (at the cost of harder to reach keys and an assymetry in the
keys (row 3 on the left side, but row 1 on the right side).

The assymetry could be solved by moving the Compose key (Caps-Lock on
the traditional keyboards position) and/or the layer shift key to the
UMLAUT layer and/or the layer shift keys to the SYMBOL layer.